+++ src/Plugin/Alipay/V2/AddPayloadSignaturePlugin.php
@@ -11,7 +11,7 @@ use Yansongda\Artful\Exception\InvalidConfigException;
 use Yansongda\Artful\Exception\ServiceNotFoundException;
 use Yansongda\Artful\Logger;
 use Yansongda\Artful\Rocket;
-use Yansongda\Pay\Config\AlipayConfig;
+use Yansongda\Pay\Config\AlipayV2Config;
 use Yansongda\Pay\Traits\AlipayTrait;
 
 class AddPayloadSignaturePlugin implements PluginInterface
@@ -41,7 +41,7 @@ class AddPayloadSignaturePlugin implements PluginInterface
      */
     protected function getSign(Rocket $rocket): string
     {
-        /** @var AlipayConfig $config */
+        /** @var AlipayV2Config $config */
         $config = self::getProviderConfig('alipay', $rocket->getParams());
         $privateKey = self::getAlipayPrivateKey($config);