+++ src/Plugin/Alipay/V2/StartPlugin.php
@@ -12,7 +12,7 @@ use Yansongda\Artful\Exception\ServiceNotFoundException;
 use Yansongda\Artful\Logger;
 use Yansongda\Artful\Rocket;
 use Yansongda\Pay\CertManager;
-use Yansongda\Pay\Config\AlipayConfig;
+use Yansongda\Pay\Config\AlipayV2Config;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Traits\AlipayTrait;
 
@@ -47,7 +47,7 @@ class StartPlugin implements PluginInterface
      */
     protected function getPayload(array $params): array
     {
-        /** @var AlipayConfig $config */
+        /** @var AlipayV2Config $config */
         $config = self::getProviderConfig('alipay', $params);
 
         return [
@@ -71,7 +71,7 @@ class StartPlugin implements PluginInterface
     /**
      * @param array<string, mixed> $params
      */
-    protected function getReturnUrl(array $params, AlipayConfig $config): string
+    protected function getReturnUrl(array $params, AlipayV2Config $config): string
     {
         if (!empty($params['_return_url'])) {
             return $params['_return_url'];
@@ -83,7 +83,7 @@ class StartPlugin implements PluginInterface
     /**
      * @param array<string, mixed> $params
      */
-    protected function getNotifyUrl(array $params, AlipayConfig $config): string
+    protected function getNotifyUrl(array $params, AlipayV2Config $config): string
     {
         if (!empty($params['_notify_url'])) {
             return $params['_notify_url'];
@@ -95,7 +95,7 @@ class StartPlugin implements PluginInterface
     /**
      * @param array<string, mixed> $params
      */
-    protected function getAppAuthToken(array $params, AlipayConfig $config): string
+    protected function getAppAuthToken(array $params, AlipayV2Config $config): string
     {
         if (!empty($params['_app_auth_token'])) {
             return $params['_app_auth_token'];
@@ -107,7 +107,7 @@ class StartPlugin implements PluginInterface
     /**
      * @throws InvalidConfigException 缺少证书配置或证书解析失败
      */
-    protected function getAppCertSn(AlipayConfig $config): string
+    protected function getAppCertSn(AlipayV2Config $config): string
     {
         $path = $config->getAppPublicCertPath();
 
@@ -121,7 +121,7 @@ class StartPlugin implements PluginInterface
     /**
      * @throws InvalidConfigException 缺少证书配置或证书解析失败
      */
-    protected function getAlipayRootCertSn(AlipayConfig $config): string
+    protected function getAlipayRootCertSn(AlipayV2Config $config): string
     {
         $path = $config->getAlipayRootCertPath();