+++ src/Plugin/Alipay/V3/AddPayloadSignaturePlugin.php
@@ -11,7 +11,7 @@ use Yansongda\Artful\Exception\InvalidConfigException;
 use Yansongda\Artful\Exception\ServiceNotFoundException;
 use Yansongda\Artful\Logger;
 use Yansongda\Artful\Rocket;
-use Yansongda\Pay\Config\AlipayConfig;
+use Yansongda\Pay\Config\AlipayV3Config;
 use Yansongda\Pay\Traits\AlipayTrait;
 use Yansongda\Supports\Collection;
 
@@ -37,7 +37,7 @@ class AddPayloadSignaturePlugin implements PluginInterface
 
         $params = $rocket->getParams();
 
-        /** @var AlipayConfig $config */
+        /** @var AlipayV3Config $config */
         $config = self::getProviderConfig('alipay', $params);
         $payload = $rocket->getPayload();
 
@@ -65,7 +65,7 @@ class AddPayloadSignaturePlugin implements PluginInterface
     /**
      * 提取签名组串中的 requestUri（path+query，不含 host）.
      */
-    protected function getSignatureUri(AlipayConfig $config, ?Collection $payload): string
+    protected function getSignatureUri(AlipayV3Config $config, ?Collection $payload): string
     {
         $url = self::getAlipayV3Url($config, $payload);
         $urlPath = parse_url($url, PHP_URL_PATH);
@@ -77,7 +77,7 @@ class AddPayloadSignaturePlugin implements PluginInterface
     /**
      * @param array<string, mixed> $params
      */
-    protected function getAppAuthToken(array $params, AlipayConfig $config): string
+    protected function getAppAuthToken(array $params, AlipayV3Config $config): string
     {
         if (!empty($params['_app_auth_token'])) {
             return $params['_app_auth_token'];